Fix catalogue tracking
The function that fails when using catalogue tracking is
ac.get_source_ra_dec(). This function seems to be broken since the
port of ata_control to the REST API. However, this function is
only used here to check if the source is up, so we can disable it
and assume that the source is up.

I have tested that this commit fixes catalogue tracking for the
trackscan and on-off modes.